+++ ai-agent/ai_agent.cpp
@@ -27,85 +27,31 @@
 
 #include "ai-agent/ai_agent.hpp"
 
-#include <sstream>
 #include <algorithm>
-#include <limits>
 
-#include "error_macros.hpp"
 #include "ai-agent/utils/utils.hpp"
-#include "ai-agent/fitness.hpp"
 #include "generator/visual_shader_generator.hpp"
+#include "ai-agent/fitness.hpp"
 
-AIAgentWorker::AIAgentWorker(ShaderGenSharedMemory* shared_memory) : start_requested(false),
-                                                                     exit_requested(false),
-                                                                     stop_requested(false),
-                                                                     maximum_population_size(0),
-                                                                     mutation_probability(0.0f), 
-                                                                     crossover_probability(0.0f), 
-                                                                     elitism_ratio(0.0f), 
-                                                                     maximum_iterations(0),
-                                                                     matching_type(MatchingType::PARAMETERS_ONLY),
-                                                                     shared_memory(shared_memory) {
-    worker = std::thread(&AIAgentWorker::worker_main, this);
-}
+namespace ai_agent_main {
 
-AIAgentWorker::~AIAgentWorker() {
-    stop_thread();
-    if (worker.joinable()) worker.join();
-}
+    inline static std::vector<std::string> generate_population(const MatchingType& matching_type, const std::string& graph, const int& maximum_population_size) noexcept;
 
-void AIAgentWorker::start_matching() {
-    {
-        std::lock_guard<std::mutex> lock(mtx);
-        start_requested = true;
-    }
-    cv.notify_one();
-}
-
-void AIAgentWorker::stop_matching() {
-    {
-        std::lock_guard<std::mutex> lock(mtx);
-        stop_requested = true;
-    }
-}
-
-void AIAgentWorker::worker_main() {
-    std::unique_lock<std::mutex> lock(mtx);
-
-    CHECK_PARAM_NULLPTR(shared_memory, "Shared memory is not set");
-
-    while (true) {
-        // Set is_stopped to false
-        shared_memory->set_is_stopped(false);
-
-        // Wait for either start command or exit request
-        cv.wait(lock, [this]() {
-            return start_requested.load() || exit_requested;
-        });
-
-        // Exit condition check
-        if (exit_requested) break;
-
-        // Reset start request
-        if (start_requested.load()) start_requested.store(false);
-
-        // It doesn't make sense to stop before even starting
-        if (stop_requested.load()) stop_requested.store(false);
-
-        CONTINUE_IF_TRUE(target_image.isNull(), "Target image is not set");
-
-        // Generate initial population
-        const std::string encoded_graph = shared_memory->get_encoded_graph();
-
-        std::vector<std::string> population = ai_agent_utils::generate_population(matching_type, encoded_graph, maximum_population_size);
+    bool genetic_algorithm(
+        const MatchingType& matching_type,
+        const std::string& encoded_graph,
+        const uint32_t* target_image_pixels,
+        const int& maximum_population_size
+    ) noexcept {
+        std::vector<std::string> population = generate_population(matching_type, encoded_graph, maximum_population_size);
 
         ImageExtractor image_extractor;
-        CONTINUE_IF_TRUE(!image_extractor.initialize(), "Failed to initialize image extractor");
+        CHECK_CONDITION_TRUE_NON_VOID(!image_extractor.initialize(), false, "Failed to initialize image extractor");
 
         // Calculate fitness
         std::vector<unsigned long> fitness_values;
         fitness_values.resize(maximum_population_size);
-        for (int i {0}; i < maximum_population_size; i++) fitness_values.at(i) = get_fitness_value(population.at(i), image_extractor);
+        for (int i {0}; i < maximum_population_size; i++) fitness_values.at(i) = get_fitness_value(population.at(i), image_extractor, target_image_pixels, 256, 256);
 
         // Create a vector of pairs of population and fitness values
         std::vector<std::pair<std::string, unsigned long>> population_fitness;
@@ -117,85 +63,115 @@ void AIAgentWorker::worker_main() {
             return a.second < b.second; // Ascending order
         });
 
-        while (maximum_iterations-- > 0) {
-            
+        return true;
+    }
+
+    unsigned long get_fitness_value(const std::string& encoded_graph, ImageExtractor& extractor, const uint32_t* target_image_pixels, const int& width, const int& height) {
+        std::string code;
+      
+        bool result{shadergen_visual_shader_generator::generate_shader(
+          shadergen_visual_shader_generator::to_proto_nodes(encoded_graph),
+          shadergen_visual_shader_generator::to_generators(encoded_graph), 
+          shadergen_visual_shader_generator::to_port_type_generators(encoded_graph),
+          shadergen_visual_shader_generator::to_input_output_connections_by_key(encoded_graph), code)};
+        CHECK_CONDITION_TRUE_NON_VOID(!result, std::numeric_limits<unsigned long>::max(), "Failed to generate shader code");
+
+        const uint32_t* current_image_pixels = extractor.render(code);
+
+        CHECK_PARAM_NULLPTR_NON_VOID(current_image_pixels, std::numeric_limits<unsigned long>::max(), "Failed to render image");
+      
+        return ai_agent_fitness::calculate_fitness(current_image_pixels, target_image_pixels, width, height);
+    }
+
+    /**
+     * @brief 
+     * 
+     * @note Josh — 10/02/2025 20:36 "my recommendation for "how can we possibly guess 
+     *       at a topology other than pure random" would be K-means clustering based on 
+     *       fourier values"
+     * 
+     * @note Josh — 07/03/2025 21:26 "I think those are good heuristics, but what I would 
+     *       probably do is start with a handful of topologies or even a few random nodes 
+     *       of each kernel type, then just choose the ones that are the closest based on 
+     *       that Fourier analysis"
+     * 
+     * @return std::vector<std::string>
+     */
+    inline static std::vector<std::string> generate_population(const MatchingType& matching_type, const std::string& graph, const int& maximum_population_size) noexcept {
+        std::vector<std::string> population;
+        population.resize(maximum_population_size);
+
+        switch (matching_type) {
+            case MatchingType::PARAMETERS_ONLY: {
+                const std::vector<std::string> entities{ai_agent_utils::split_string(graph, ',')};
+                for (const auto& entity : entities) {
+                    const std::vector<std::string> tokens{ai_agent_utils::split_string(entity, ';')};
+                    const int entity_type = std::stoi(tokens.at(0));
+                    if (entity_type == 0) {
+                        // Node
+                        // Format: entity_type;node_id;node_type;field_number=value;field_number=value;...
+                        const int n_id = std::stoi(tokens.at(1));
+                        const int oneof_value_field_number = std::stoi(tokens.at(2));
+                        const std::vector<std::string> parameters{tokens.begin() + 3, tokens.end()};
+                        for (int i{0}; i < maximum_population_size; ++i) {
+                            std::vector<std::string> new_parameters;
+                            new_parameters.resize(parameters.size());
+                            for (int j{0}; j < parameters.size(); ++j) {
+                                const std::string parameter{parameters.at(j)}; // Format: field_number=value
+                                const std::vector<std::string> parameter_tokens{ai_agent_utils::split_string(parameter, '=')};
+                                const int field_number = std::stoi(parameter_tokens.at(0));
+                                const ai_agent_utils::DiscreteContinuousRangeVariant range{ai_agent_utils::get_range_for_field(oneof_value_field_number, field_number)};
+                                std::string new_parameter;
+                                new_parameter += std::to_string(field_number) + '=';
+                                if (std::holds_alternative<std::vector<int>>(range)) {
+                                    const std::vector<int> discrete_range{std::get<std::vector<int>>(range)};
+                                    const int random_index{ai_agent_utils::random_int_inclusive<int>(0, (int)discrete_range.size() - 1)};
+                                    new_parameter += std::to_string(discrete_range.at(random_index)) + ';';
+                                } else if (std::holds_alternative<std::pair<ai_agent_utils::ContinuousRangeVariant, ai_agent_utils::ContinuousRangeVariant>>(range)) {
+                                    const std::pair<ai_agent_utils::ContinuousRangeVariant, ai_agent_utils::ContinuousRangeVariant> continuous_range{std::get<std::pair<ai_agent_utils::ContinuousRangeVariant, ai_agent_utils::ContinuousRangeVariant>>(range)};
+                                    if (std::holds_alternative<int>(continuous_range.first) && std::holds_alternative<int>(continuous_range.second)) {
+                                        const int lower_bound{std::get<int>(continuous_range.first)};
+                                        const int upper_bound{std::get<int>(continuous_range.second)};
+                                        const float random_value{ai_agent_utils::random_real_inclusive<float>(lower_bound, upper_bound)};
+                                        new_parameter += std::to_string(random_value) + ';';
+                                    } else if (std::holds_alternative<float>(continuous_range.first) && std::holds_alternative<float>(continuous_range.second)) {
+                                        const float lower_bound{std::get<float>(continuous_range.first)};
+                                        const float upper_bound{std::get<float>(continuous_range.second)};
+                                        const float random_value{ai_agent_utils::random_real_inclusive<float>(lower_bound, upper_bound)};
+                                        new_parameter += std::to_string(random_value) + ';';
+                                    }
+                                } else {
+                                    new_parameter += "0;";
+                                }
+                                new_parameter.pop_back(); // Remove the last semicolon
+                                new_parameters.at(j) = new_parameter;
+                            }
+                            std::string encoded_node;
+                            encoded_node += std::to_string(entity_type) + ';';
+                            encoded_node += std::to_string(n_id) + ';';
+                            encoded_node += std::to_string(oneof_value_field_number) + ';';
+                            for (const auto& new_parameter : new_parameters) encoded_node += new_parameter + ';';
+
+                            encoded_node.pop_back(); // Remove the last semicolon
+
+                            population.at(i) += encoded_node + ',';
+                        }
+                    } else {
+                        // Connection
+                        for (int i{0}; i < maximum_population_size; ++i) population.at(i) += entity + ',';
+                    }
+                }
+
+                for (auto& entity : population) {
+                    if (!entity.empty()) entity.pop_back(); // Remove the last comma
+                }
+                break;
+            }
+            default:
+                break;
         }
 
-        // Process all pending requests
-        // while (process_counter > 0) {
-        //     process_counter--;
-        //     lock.unlock();
-
-        //     // Process with interrupt checks
-        //     bool completed = false;
-        //     auto start = std::chrono::steady_clock::now();
-            
-        //     // Work simulation with interrupt checks
-        //     while (!completed) {
-        //         // Do chunk of work
-        //         DEBUG_PRINT("Processing...");
-                
-        //         // Check for stop every 100ms
-        //         std::this_thread::sleep_for(std::chrono::milliseconds(100));
-                
-        //         // Check completion condition
-        //         auto now = std::chrono::steady_clock::now();
-        //         if (now - start > std::chrono::milliseconds(500)) {
-        //             completed = true;
-        //         }
-
-        //         // Check for stop request
-        //         if (stop_requested.load()) {
-        //             DEBUG_PRINT("Interrupting current work");
-        //             completed = true;
-        //         }
-        //     }
-
-        //     lock.lock();
-            
-        //     // Exit processing loop if stop requested
-        //     if (stop_requested.load()) {
-        //         stop_requested.store(false);
-        //         process_counter = 0; // Clear remaining tasks
-        //         break;
-        //     }
-        // }
+        return population;
     }
-    DEBUG_PRINT("Worker thread exiting cleanly");
-}
 
-void AIAgentWorker::stop_thread() {
-    {
-        std::lock_guard<std::mutex> lock(mtx);
-        exit_requested = true;
-    }
-    cv.notify_one();
-}
-
-unsigned long AIAgentWorker::get_fitness_value(const std::string& encoded_graph, ImageExtractor& extractor) {
-  std::string code;
-
-  bool result{shadergen_visual_shader_generator::generate_shader(
-    shadergen_visual_shader_generator::to_proto_nodes(encoded_graph),
-    shadergen_visual_shader_generator::to_generators(encoded_graph), 
-    shadergen_visual_shader_generator::to_port_type_generators(encoded_graph),
-    shadergen_visual_shader_generator::to_input_output_connections_by_key(encoded_graph), code)};
-  CHECK_CONDITION_TRUE_NON_VOID(!result, std::numeric_limits<unsigned long>::max(), "Failed to generate shader code");
-
-  CHECK_CONDITION_TRUE_NON_VOID(target_image.isNull(), std::numeric_limits<unsigned long>::max(), "No target image loaded");
-
-  QImage current_image = extractor.render(code);
-
-  CHECK_CONDITION_TRUE_NON_VOID(current_image.isNull(), std::numeric_limits<unsigned long>::max(), "Failed to render image");
-  CHECK_CONDITION_TRUE_NON_VOID(current_image.size() != target_image.size(), std::numeric_limits<unsigned long>::max(), "Size mismatch");
-
-  // Retrieve pointers to the pixel data.
-  // QImage::bits() returns a pointer to the first pixel, and since our format is ARGB32,
-  // we can safely reinterpret_cast to a uint32_t pointer.
-  const uint32_t* pixels1 = reinterpret_cast<const uint32_t*>(current_image.bits());
-  const uint32_t* pixels2 = reinterpret_cast<const uint32_t*>(target_image.bits());
-  int width = current_image.width();
-  int height = current_image.height();
-
-  return ai_agent_fitness::calculate_fitness(pixels1, pixels2, width, height);
-}
+} // namespace ai_agent_main
