+++ gui/controller/ai_agent_worker.hpp
@@ -25,35 +25,61 @@
 /*                                                                               */
 /*********************************************************************************/
 
-#ifndef SHADER_GEN_AI_AGENT_IMAGE_EXTRACTOR_HPP
-#define SHADER_GEN_AI_AGENT_IMAGE_EXTRACTOR_HPP
+#ifndef AI_AGENT_WORKER_HPP
+#define AI_AGENT_WORKER_HPP
 
-#include <QtOpenGL/QOpenGLFunctions_4_3_Core>
-#include <QOffscreenSurface>
-#include <QOpenGLContext>
-#include <QOpenGLFramebufferObject>
-#include <QtOpenGL/QOpenGLShaderProgram>
-#include <memory>
-#include <string>
+#include <thread>
+#include <mutex>
+#include <condition_variable>
+#include <chrono>
 
-class ImageExtractor : protected QOpenGLFunctions_4_3_Core {
-  public:
-  ImageExtractor();
-  ~ImageExtractor();
+#include <QImage>
 
-  bool initialize();
-  const uint32_t* render(const std::string& code);
+#include "ai-agent/ai_agent.hpp"
+#include "ai-agent/shared_memory.hpp"
+#include "ai-agent/utils/image_extractor.hpp"
 
+class AIAgentWorker {
+public:
+    AIAgentWorker(ShaderGenSharedMemory* shared_memory);
+    ~AIAgentWorker();
+
+    void start_matching();
+    void stop_matching();
+
+    void worker_main();
+
+    void set_maximum_population_size(const int& maximum_population_size) { this->maximum_population_size = maximum_population_size; }
+    void set_mutation_probability(const float& mutation_probability) { this->mutation_probability = mutation_probability; }
+    void set_crossover_probability(const float& crossover_probability) { this->crossover_probability = crossover_probability; }
+    void set_elitism_ratio(const float& elitism_ratio) { this->elitism_ratio = elitism_ratio; }
+    void set_maximum_iterations(const int& maximum_iterations) { this->maximum_iterations = maximum_iterations; }
+
+    void set_matching_type(const ai_agent_main::MatchingType& matching_type) { this->matching_type = matching_type; }
+    void set_target_image(const QImage& target_image) { this->target_image = target_image; }
+    
 private:
-  void init_buffers();
-  bool compile_shader(const std::string& code, QOpenGLShaderProgram& program);
+    std::thread worker;
+    
+    std::mutex mtx;
+    std::condition_variable cv;
+    int process_counter;
+    std::atomic<bool> start_requested;
+    std::atomic<bool> exit_requested;
+    std::atomic<bool> stop_requested;
 
-  QOpenGLContext* context;
-  QOffscreenSurface* surface;
-  QOpenGLFramebufferObject* fbo;
-  
-  GLuint VAO;
-  GLuint VBO;
+    int maximum_population_size;
+    float mutation_probability;
+    float crossover_probability;
+    float elitism_ratio;
+    int maximum_iterations;
+
+    ai_agent_main::MatchingType matching_type;
+    QImage target_image;
+
+    ShaderGenSharedMemory* shared_memory;
+
+    void stop_thread();
 };
 
-#endif  // SHADER_GEN_AI_AGENT_IMAGE_EXTRACTOR_HPP
+#endif // AI_AGENT_WORKER_HPP
